refactor: remove useless service and rename models
